// src/lib/generated/prisma/inputTypeSchemas/NestedFloatFilterSchema.ts
import type { Prisma } from '@prisma/client';
import { z } from 'zod';

export const NestedFloatFilterSchema: z.ZodType<Prisma.NestedFloatFilter> = z
  .object({
    equals: z.number().optional(),
    in: z.union([z.number().array(), z.number()]).optional(),
    notIn: z.union([z.number().array(), z.number()]).optional(),
    lt: z.number().optional(),
    lte: z.number().optional(),
    gt: z.number().optional(),
    gte: z.number().optional(),
    not: z.union([z.number(), z.lazy(() => NestedFloatFilterSchema)]).optional()
  })
  .strict();

export default NestedFloatFilterSchema;